refactor: Major refactor in internals and module layout.
This includes a large refactor of the layout of the filesystem namely:
- All futures have been separated out to their own modules, and made to
be generic over the underlying IO device.
- Refactored how the io_uring handles completions, and moved completion
handling to the futures themselves, this allows flexibility and to keep
the ring code as clean as possible.
- Fully handles address management for all sockets now.
- Adds in oneshot + multishot result handlers for futures to use, these
are currently simple implementations but allows for an abstraction for
easy changes later.