feat(buffers): This is a large change that moves libuio to owned buffers
This change set is required because io_uring under the hood is
completion based, as opposed to intent based like epoll/kqueue/etc this
means that the io_uring subsystem actually is what is reading/writing to
the buffers specified to calls like recv/send.

Without io_uring owning the buffers in use, there is a possibility that
the future that kicked off the I/O is dropped (including the buffer)
while io_uring is writing/reading to/from it. This would lead to very
very bad things and undefined behavior. Since we currently can't express
the lifetime constraints we really need via the Rust type system. The
only way to guarantee soundness here, is to own the buffers in question
end to end.

With the above said, this isn't all that big of a deal in the end, the
system always returns the buffers in use, and the ergonomics can be
greatly increased with proper helpers. Also this paves the way for
Recv[Msg]Multi setup and provided buffers for further enhancements.